refactors api and improves naming of file structure
