+++ api/groa_ds_api/test_api.py
@@ -1,5 +1,5 @@
 from fastapi.testclient import TestClient
-from apps import create_app
+from groa_ds_api import create_app
 
 application = create_app()
 client = TestClient(application)
@@ -23,10 +23,10 @@ def test_decades():
 def test_recommendations():
     payload = {
         "user_id": 11,
-        "number_of_recommendations": 50,
+        "num_recs": 50,
         "good_threshold": 5,
         "bad_threshold": 4,
         "harshness": 1
     }
-    response = client.post("/recommendation", json=payload)
+    response = client.post("/recommendations", json=payload)
     assert response.status_code == 200
